# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Smart WiFi Pass MVP
  description: PDF 축소 명세의 FastAPI 단일 백엔드입니다.
  version: 1.0.0
paths:
...
  /public/order-claims/exchange:
    post:
      tags:
      - Public
      summary: Exchange Claim
      operationId: exchange_claim_public_order_claims_exchange_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClaimExchangeRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /public/otp/send:
    post:
      tags:
      - Public
      summary: Send Otp
      operationId: send_otp_public_otp_send_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OtpSendRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /public/otp/confirm:
    post:
      tags:
      - Public
      summary: Confirm Otp
      operationId: confirm_otp_public_otp_confirm_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OtpConfirmRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /public/passes/{pass_id}/activate:
    post:
      tags:
      - Public
      summary: Activate Pass
      operationId: activate_pass_public_passes__pass_id__activate_post
      parameters:
      - name: pass_id
        in: path
        required: true
        schema:
          type: string
          title: Pass Id
      - name: x-portal-session
        in: header
        required: true
        schema:
          type: string
          title: X-Portal-Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /public/passes/{pass_id}:
    get:
      tags:
      - Public
      summary: Get Pass
      operationId: get_pass_public_passes__pass_id__get
      parameters:
      - name: pass_id
        in: path
        required: true
        schema:
          type: string
          title: Pass Id
      - name: x-portal-session
        in: header
        required: true
        schema:
          type: string
          title: X-Portal-Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /public/kiosk/upsell-hint:
    get:
      tags:
      - Public
      summary: Upsell Hint
      operationId: upsell_hint_public_kiosk_upsell_hint_get
      parameters:
      - name: x-portal-session
        in: header
        required: true
        schema:
          type: string
          title: X-Portal-Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /public/upsell-hint:
    get:
      tags:
      - Public
      summary: Upsell Hint
      operationId: upsell_hint_public_upsell_hint_get
      parameters:
      - name: x-portal-session
        in: header
        required: true
        schema:
          type: string
          title: X-Portal-Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /public/rewards/grants/{grant_id}/options:
    get:
      tags:
      - Public
      summary: Reward Options
      operationId: reward_options_public_rewards_grants__grant_id__options_get
      parameters:
      - name: grant_id
        in: path
        required: true
        schema:
          type: string
          title: Grant Id
      - name: x-portal-session
        in: header
        required: true
        schema:
          type: string
          title: X-Portal-Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /public/rewards/grants/{grant_id}/choose:
    post:
      tags:
      - Public
      summary: Choose Reward
      operationId: choose_reward_public_rewards_grants__grant_id__choose_post
      parameters:
      - name: grant_id
        in: path
        required: true
        schema:
          type: string
          title: Grant Id
      - name: x-portal-session
        in: header
        required: true
        schema:
          type: string
          title: X-Portal-Session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RewardChooseRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /public/rewards/{grant_id}/choose:
    post:
      tags:
      - Public
      summary: Choose Reward
      operationId: choose_reward_public_rewards__grant_id__choose_post
      parameters:
      - name: grant_id
        in: path
        required: true
        schema:
          type: string
          title: Grant Id
      - name: x-portal-session
        in: header
        required: true
        schema:
          type: string
          title: X-Portal-Session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RewardChooseRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /public/coupons:
    get:
      tags:
      - Public
      summary: List Coupons
      operationId: list_coupons_public_coupons_get
      parameters:
      - name: x-portal-session
        in: header
        required: true
        schema:
          type: string
          title: X-Portal-Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /public/stores/{store_id}/privacy-notice:
    get:
      tags:
      - Public
      summary: Privacy Notice
      operationId: privacy_notice_public_stores__store_id__privacy_notice_get
      parameters:
      - name: store_id
        in: path
        required: true
        schema:
          type: string
          title: Store Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /public/coupons/{coupon_id}/redeem:
    post:
      tags:
      - Public
      summary: Redeem Coupon
      operationId: redeem_coupon_public_coupons__coupon_id__redeem_post
      parameters:
      - name: coupon_id
        in: path
        required: true
        schema:
          type: string
          title: Coupon Id
      - name: x-portal-session
        in: header
        required: true
        schema:
          type: string
          title: X-Portal-Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ClaimExchangeRequest:
      properties:
        orderClaim:
          type: string
          title: Orderclaim
      type: object
      required:
      - orderClaim
      title: ClaimExchangeRequest
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    OtpConfirmRequest:
      properties:
        challengeId:
          type: string
          title: Challengeid
        code:
          type: string
          pattern: ^\d{6}$
          title: Code
      type: object
      required:
      - challengeId
      - code
      title: OtpConfirmRequest
    OtpSendRequest:
      properties:
        verificationTicket:
          type: string
          title: Verificationticket
        phone:
          type: string
          title: Phone
      type: object
      required:
      - verificationTicket
      - phone
      title: OtpSendRequest
...
    RewardChooseRequest:
      properties:
        benefitId:
          type: string
          title: Benefitid
        fulfillMode:
          type: string
          pattern: ^(IMMEDIATE|COUPON_7D)$
          title: Fulfillmode
        orderId:
          anyOf:
          - type: string
          - type: 'null'
          title: Orderid
      type: object
      required:
      - benefitId
      - fulfillMode
      title: RewardChooseRequest
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...